# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Evaluation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List approved projects
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, votes, featured]
            default: featured
      responses:
        "200":
          description: Paginated project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProjects"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/approve:
    put:
      operationId: approveProject
      tags: [projects]
      summary: Approve a project (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/reject:
    put:
      operationId: rejectProject
      tags: [projects]
      summary: Reject a project (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/vote:
    post:
      operationId: voteProject
      tags: [votes]
      summary: Vote on a project (once per user)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote recorded or removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/my-vote:
    get:
      operationId: getMyVote
      tags: [votes]
      summary: Check if the current user voted on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyVoteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/stats/top-voted:
    get:
      operationId: getTopVotedProjects
      tags: [projects]
      summary: Get top voted projects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top voted projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        voteCount:
          type: integer
        userId:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - status
        - isFeatured
        - imageUrl
        - voteCount
        - userId
        - ownerName
        - createdAt
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        voteCount:
          type: integer
        userId:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - status
        - isFeatured
        - imageUrl
        - voteCount
        - userId
        - ownerName
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - description
        - category
    PaginatedProjects:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - projects
        - total
        - page
        - limit
        - totalPages
...
    VoteResponse:
      type: object
      properties:
        voted:
          type: boolean
        voteCount:
          type: integer
      required:
        - voted
        - voteCount
    MyVoteResponse:
      type: object
      properties:
        voted:
          type: boolean
      required:
        - voted
...